refactor(contacts): extract PropertyFormField to remove duplication
